feat(api): move sync client to its own module
