+++ courtlistener/sync_client/__init__.py
@@ -1,0 +1,13 @@
+from courtlistener.sync_client.alerts import DocketAlerts, SearchAlerts
+from courtlistener.sync_client.citation_lookup import CitationLookup
+from courtlistener.sync_client.client import CourtListener
+from courtlistener.sync_client.resource import Resource, ResourceIterator
+
+__all__ = [
+    "CitationLookup",
+    "CourtListener",
+    "DocketAlerts",
+    "Resource",
+    "ResourceIterator",
+    "SearchAlerts",
+]
